Integrate link-sharing into dashboard
