+++ packages/frontend/src/pages/routes/UnauthenticatedRoutes.tsx
@@ -3,16 +3,15 @@ import { Routes, Route, Navigate } from 'react-router-dom';
 
 import LandingPage from '../public/LandingPage';
 import SignInPage from '../public/SignInPage';
-import HomePage from '../public/HomePage';
 
 interface UnauthenticatedRoutesProps {}
 
 const UnauthenticatedRoutes: React.FC<UnauthenticatedRoutesProps> = () => {
   return (
     <Routes>
-      <Route path="auth" element={<SignInPage />} />
-      <Route path="landing" element={<LandingPage />} />
-      <Route path="*" element={<Navigate to="/landing" replace />} />
+      <Route path="/" element={<LandingPage />} />
+      <Route path="/auth" element={<SignInPage />} />
+      <Route path="*" element={<Navigate to="/auth" replace />} />
     </Routes>
   );
 };
